feat: Created separate folder for modals and toasts and related changes
