+++ stockmonitor/frontend/src/components/HomePage.tsx
@@ -10,7 +10,8 @@ import api from '../api.tsx';
 import { Stock, Additem, AddStockProps } from '../interfaces/index.tsx';
 import StockInfoCard from './StockInfoCard.tsx';
 import NavBar from './Navbar.tsx';
-import AddStockModal from './AddStockModal.tsx';
+import AddStockModal from './modals/AddStockModal.tsx';
+import AddStockComponent from './AddStockTop.tsx';
 
 
 const HomePage: React.FC = () => {
@@ -139,82 +140,3 @@ const StockWishlist = ({ stocks, isLoading }: { stocks: Stock[], isLoading: bool
         </Container>
     )
 }
-
-const AddStockComponent: React.FC<AddStockProps> = (
-    { stocks, addStock, handleChange, handleSubmit, showAddStockModal }
-) => {
-
-    const [error, setError] = useState("");
-    const checkDuplicateStock = (): boolean => {
-        for (const stock of stocks) {
-            if (stock.symbol.trim() === addStock.symbol.trim()) {
-                setError('Stock already exists in your wishlist');
-                return true;
-            }
-        }
-        setError("");
-        return false;
-    }
-
-    useEffect(() => {
-        checkDuplicateStock();
-    }, [addStock]);
-
-    return (
-        <Box>
-            <Box
-                component="form"
-                onSubmit={(e) => {
-                    e.preventDefault();
-                    if (checkDuplicateStock()) {
-                        return;
-                    };
-                    handleSubmit(e);
-                }}
-                sx={{
-                    display: {
-                        xs: "none",
-                        md: "flex",
-                    },
-                    alignItems: 'center',
-                    padding: "8px 4px",
-                }}
-            >
-                <TextField
-                    size='small'
-                    id="outlined-size-small"
-                    sx={{ borderBottom: "0" }}
-                    InputProps={{
-                        style: {
-                            backgroundColor: "white",
-                            width: "35ch",
-                        }
-                    }}
-                    label={error.length == 0 ? 'Add Stock': error}
-                    variant="filled"
-                    name="symbol"
-                    value={addStock.symbol}
-                    onChange={handleChange}
-                    type='text'
-                    required
-                    error={error.length !== 0}
-                />
-
-                <Button
-                    type="submit"
-                    color="success"
-                    variant="contained"
-                    sx={{ margin: '0 12px' }}
-                    name='submit'
-                >
-                    Submit
-                </Button>
-            </Box>
-
-            {/* Icon to show modal for adding stock on smaller screens */}
-            <Box px={2} onClick={showAddStockModal} sx={{display:{ sm: 'block' , md: 'none'}}}><AddIcon /></Box>
-        </Box>
-    )
-}
-
-
