+++ stockmonitor/frontend/src/components/modals/DeleteStockModal.tsx
