+++ stockmonitor/frontend/src/components/toast/ErrorDisplay.tsx
